Tested to result in a proper database
Test manually by POST the following

{
  "end_user": "Player1",
  "flavour": "m1.xlarge",
  "job_hash": "test",
  "extra_files":  { "/etc/init.d/something_important" : "jkajkggk
wegnkljkorgjklrgnklerklrg", "field2" : "value2"},
  "container_image_descriptor": "{ \"field1\" : \"value1\", \"field2\" :
\"value2\"}",
  "container_runtime_descriptor": "{ \"field1\" : \"value1\", \"field2\"
: \"value2\"}"
}

+++ consonance-webservice/src/main/java/io/consonance/webservice/jdbi/JobDAO.java
@@ -31,11 +31,11 @@ public class JobDAO extends AbstractDAO<Job> {
         super(factory);
     }
 
-    public Job findById(long id) {
+    public Job findById(int id) {
         return get(id);
     }
 
-    public long create(Job job) {
+    public int create(Job job) {
         return persist(job).getJobId();
     }
